Pull out initial transfer into a worker (#597)
