+++ pokemongo_bot/cell_workers/__init__.py
@@ -3,3 +3,4 @@
 from pokemon_catch_worker import PokemonCatchWorker
 from seen_fort_worker import SeenFortWorker
 from move_to_fort_worker import MoveToFortWorker
+from initial_transfer_worker import InitialTransferWorker
